move constant files under c/lib to c/constants
