+++ packages/client/constants/sampleHomepageCode.ts
