+++ packages/client/pages/orbital/proposal.mdx
@@ -1,7 +1,7 @@
-import authors from "@client/lib/authors";
-
 import BlogImage from "@client/components/blogImage";
 import BlogLayout from "@client/layouts/blog";
+
+import authors from "@client/constants/authors";
 
 export default function ({ children }) {
   return (
